db data added to renderer
