+++ css/style.css
@@ -72,6 +72,7 @@ html, body {
 
 #todolist {
 	background: #FF6666;
+	scroll-behavior: auto;
 	color: #FFF;
 	padding: 2rem 3rem 3rem;
 	box-shadow: -20px -20px 0px 0px rgba(100,100,100,.1);
