+++ pages/viewgames.html
@@ -11,11 +11,12 @@
         <ul>
             <a href="#">
             <li class="" style="">
-                <span class="label">Code a todo list</span><span class="aboutgame">this is a good game</span>
+                <span class="label">Game 1</span><span class="aboutgame">this is a good game</span>
             </li>
             </a>
         </ul>
         
     </main>
 </body>
+<script src="renderer.js"></script>
 </html>
